Move collection radio buttons / check boxes back to FormOptionsHelper
[Carlos Antonio da Silva + Rafael Mendonça França]
